+++ SQLiteStudio3/guiSQLiteStudio/dialogs/dbdialog.cpp
@@ -575,6 +575,13 @@ bool DbDialog::validate()
         setValidState(ui->fileEdit, false, tr("Enter a database file path."));
         fileState = false;
     }
+    else if (QFileInfo(getPath()).isRelative())
+    {
+        setValidStateWarning(ui->fileEdit,
+            tr("You're using a relative file path, which will be resolved to \"%1\" according to the application's working directory. It's always better to use absolute file path to avoid unexpected database location.")
+                             .arg(QFileInfo(getPath()).absoluteFilePath()));
+        fileState = false;
+    }
 
     if (fileState)
     {