+++ ChangeLog.md
@@ -1,6 +1,7 @@
 # ChangeLog
 
 ### 3.4.8
+- CHANGE: #5146 If user enters a relative file path to the database dialog, a warning will issued to the user pointing out actual resolved path for the relative path entered.
 - BUGFIX: #5145 Fixed Extensions Manager crashing the app if the percentile extension was loaded before. The percentile SQLite extension became corrupted in SQLite 3.46.x, then it got fixed, but the fix is not included in SQLite 3.47.1 source package.
 - BUGFIX: #5144 Fixed Extensions Manager commit button when all extensions got removed from the list.